refactor: 전체보기 버튼 컴포넌트 추가
