+++ src/components/Common/index.ts
@@ -26,3 +26,4 @@ export { default as Badge } from './Badge/Badge';
 export { default as WriteButton } from './WriteButton/WriteButton';
 export { default as Text } from './Text/Text';
 export { default as TopBar } from './TopBar/TopBar';
+export { default as ShowAllButton } from './ShowAllButton/ShowAllButton';
